Change double for float to speed up the simulation

+++ 2DLBM/LBMCell.h
@@ -31,10 +31,10 @@
 
 struct LBMCell
 {
-	double mass = 1;           //cell mass
-	double rho = 1;            //cell density
-	std::array<double, 2> u;   //cell velocity
-	std::array<double, 9> f;   //particle distribution functions
+	float mass = 1;           //cell mass
+	float rho = 1;            //cell density
+	std::array<float, 2> u;   //cell velocity
+	std::array<float, 9> f;   //particle distribution functions
 	int tag = FLUID;           //cell type
 	int BC = INTERNAL;         //FIXEDV if the cell velocity is fixed
 	bool newInterface = false; //true if the cell will change to a interface cell in the next step
